// frontend/src/client/types.gen.ts
export type ChallengeUpdate = {
    title?: (string | null);
    author?: (string | null);
    url?: (string | null);
    description?: (string | null);
    date_posted?: (string | null);
    is_active?: (boolean | null);
};